controlled input in newpost-component
Lifted state to app, newpost is now controlled component of app. Conditional rendering of buttons in newpost. When in edit-mode, messagetext is copied into app-state. Need to now change the submit-method, so that the edited message is PUT instead of added as new post -> conditional rendering of onclick-methods on submit button.